fix: menu.html from relearn theme
Hugo searches for template files first in the /layouts/
directory.  To override a theme file, just copy it into
that dir and then modify as needed.

I mention this because the relearn theme uses a deprecated
`.site.IsMultiLingual` and we just needed to update that one call
to build correctly.

So I copied `menu.html` from the theme into our local layouts/
directory and made the change there.